añadir pantalla consumo y arreglar errores

+++ app/src/main/java/com/example/pde_consumo_energetico/Consumo.kt
@@ -1,5 +1,5 @@
 package com.example.pde_consumo_energetico
 
-data class Consumo(val semana: Int, val consumo: Int){
-    constructor() : this(0, 0)
+data class Consumo(val semana: Int, val consumo: Double){
+    constructor() : this(0, 0.0)
 }
